Create dedicated test crate and move demonlist tests there
This way, all integration tests are compiled into the same crate, which
allows them to run parallelized and reduces the number of crate that are
compiled by `cargo test`

Also long term means the test boilerplate is no longer duplicated
between *-api crates.
