+++ pointercrate-test/tests/it/demonlist/demon.rs
@@ -2,15 +2,13 @@ use pointercrate_demonlist::LIST_MODERATOR;
 use rocket::http::Status;
 use sqlx::{Pool, Postgres};
 
-mod setup;
-
 const DEFAULT_THUMBNAIL: &'static str = "https://i.ytimg.com/vi/zebrafishes/mqdefault.jpg";
 
 #[sqlx::test(migrations = "../migrations")]
 async fn test_add_demon_ratelimits(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
-    let user = setup::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
+    let user = pointercrate_test::user::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
 
     let demon = serde_json::json! {{"name": "Bloodbath", "requirement": 90, "position": 1, "verifier": "Riot", "publisher": "Riot", "creators": []}};
 
@@ -34,9 +32,9 @@ async fn test_add_demon_ratelimits(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn test_default_thumbnail_no_video(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
-    let user = setup::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
+    let user = pointercrate_test::user::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
 
     let demon = serde_json::json! {{"name": "Bloodbath", "requirement": 90, "position": 1, "verifier": "Riot", "publisher": "Riot", "creators": []}};
 
@@ -55,9 +53,9 @@ async fn test_default_thumbnail_no_video(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn test_default_thumbnail_linked_banned(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
-    let user = setup::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
+    let user = pointercrate_test::user::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
 
     let demon = serde_json::json! {{"name": "Bloodbath", "requirement": 90, "position": 1, "verifier": "Riot", "publisher": "Riot", "creators": [], "video": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"}};
 
@@ -81,9 +79,9 @@ async fn test_default_thumbnail_linked_banned(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn test_default_thumbnail_with_video(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
-    let user = setup::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
+    let user = pointercrate_test::user::system_user_with_perms(LIST_MODERATOR, &mut connection).await;
 
     let demon = serde_json::json! {{"name": "Bloodbath", "requirement": 90, "position": 1, "verifier": "Riot", "publisher": "Riot", "creators": [], "video": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"}};
