+++ pointercrate-test/tests/it/demonlist/mod.rs
@@ -1,0 +1,4 @@
+mod claim;
+mod demon;
+mod player;
+mod record;
