+++ pointercrate-test/tests/it/demonlist/player.rs
@@ -5,8 +5,6 @@ use pointercrate_demonlist::{
 use rocket::http::Status;
 use sqlx::{PgConnection, Pool, Postgres};
 
-mod setup;
-
 async fn create_players(connection: &mut PgConnection) -> (DatabasePlayer, DatabasePlayer) {
     let mut banned = DatabasePlayer::by_name_or_create("stardust1971", &mut *connection).await.unwrap();
     banned.ban(&mut *connection).await.unwrap();
@@ -18,7 +16,7 @@ async fn create_players(connection: &mut PgConnection) -> (DatabasePlayer, Datab
 
 #[sqlx::test(migrations = "../migrations")]
 async fn test_unauthenticated_pagination(pool: Pool<Postgres>) {
-    let (client, mut connection) = setup::setup_rocket(pool).await;
+    let (client, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let (_, unbanned) = create_players(&mut connection).await;
 
@@ -30,10 +28,10 @@ async fn test_unauthenticated_pagination(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn test_authenticated_pagination(pool: Pool<Postgres>) {
-    let (client, mut connection) = setup::setup_rocket(pool).await;
+    let (client, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let (_, unbanned) = create_players(&mut connection).await;
-    let user = setup::add_normal_user(&mut *connection).await;
+    let user = pointercrate_test::user::add_normal_user(&mut *connection).await;
 
     let json: Vec<Player> = client
         .get("/api/v1/players")
@@ -48,10 +46,10 @@ async fn test_authenticated_pagination(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn test_list_helper_pagination(pool: Pool<Postgres>) {
-    let (client, mut connection) = setup::setup_rocket(pool).await;
+    let (client, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let (banned, unbanned) = create_players(&mut connection).await;
-    let user = setup::system_user_with_perms(LIST_HELPER, &mut *connection).await;
+    let user = pointercrate_test::user::system_user_with_perms(LIST_HELPER, &mut *connection).await;
 
     let json: Vec<Player> = client
         .get("/api/v1/players")
