+++ pointercrate-test/tests/it/demonlist/record.rs
@@ -9,11 +9,9 @@ use rocket::http::{Header, Status};
 use sqlx::{PgConnection, Pool, Postgres};
 use std::{net::IpAddr, str::FromStr};
 
-mod setup;
-
 #[sqlx::test(migrations = "../migrations")]
 async fn paginate_records_unauthorized(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let (p1, r1, r2, r3) = setup_pagination_tests(&mut connection).await;
 
@@ -29,12 +27,12 @@ async fn paginate_records_unauthorized(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn paginate_records_with_verified_claim(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let (p1, r1, r2, r3) = setup_pagination_tests(&mut connection).await;
-    let user = setup::add_normal_user(&mut connection).await;
+    let user = pointercrate_test::user::add_normal_user(&mut connection).await;
 
-    setup::put_claim(user.inner().id, p1, true, false, &mut connection).await;
+    pointercrate_test::demonlist::put_claim(user.inner().id, p1, true, false, &mut connection).await;
 
     let json: Vec<serde_json::Value> = clnt
         .get(format!("/api/v1/records/?player={}", p1))
@@ -49,12 +47,12 @@ async fn paginate_records_with_verified_claim(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn paginate_records_with_unverified_claim(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let (p1, r1, r2, r3) = setup_pagination_tests(&mut connection).await;
-    let user = setup::add_normal_user(&mut connection).await;
+    let user = pointercrate_test::user::add_normal_user(&mut connection).await;
 
-    setup::put_claim(user.inner().id, p1, false, false, &mut connection).await;
+    pointercrate_test::demonlist::put_claim(user.inner().id, p1, false, false, &mut connection).await;
 
     let json: Vec<serde_json::Value> = clnt
         .get(format!("/api/v1/records/?player={}", p1))
@@ -68,12 +66,12 @@ async fn paginate_records_with_unverified_claim(pool: Pool<Postgres>) {
 
 #[sqlx::test(migrations = "../migrations")]
 async fn paginate_records_with_verified_claim_wrong_player(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let (p1, r1, r2, r3) = setup_pagination_tests(&mut connection).await;
-    let user = setup::add_normal_user(&mut connection).await;
+    let user = pointercrate_test::user::add_normal_user(&mut connection).await;
 
-    setup::put_claim(user.inner().id, p1, true, false, &mut connection).await;
+    pointercrate_test::demonlist::put_claim(user.inner().id, p1, true, false, &mut connection).await;
 
     let json: Vec<serde_json::Value> = clnt.get("/api/v1/records/?player=2").authorize_as(&user).get_result().await;
 
@@ -84,25 +82,25 @@ async fn setup_pagination_tests(connection: &mut PgConnection) -> (i32, i32, i32
     let player1 = DatabasePlayer::by_name_or_create("stardust1971", connection).await.unwrap();
     let player2 = DatabasePlayer::by_name_or_create("stardust1972", connection).await.unwrap();
 
-    let demon1 = setup::add_demon("Bloodbath", 1, 87, player1.id, player1.id, connection).await;
-    let demon2 = setup::add_demon("Bloodlust", 2, 53, player1.id, player1.id, connection).await;
+    let demon1 = pointercrate_test::demonlist::add_demon("Bloodbath", 1, 87, player1.id, player1.id, connection).await;
+    let demon2 =  pointercrate_test::demonlist::add_demon("Bloodlust", 2, 53, player1.id, player1.id, connection).await;
 
-    let r1 = setup::add_simple_record(100, player1.id, demon1, RecordStatus::Approved, connection).await;
-    let r2 = setup::add_simple_record(70, player1.id, demon2, RecordStatus::Rejected, connection).await;
-    let r3 = setup::add_simple_record(100, player2.id, demon2, RecordStatus::Rejected, connection).await;
+    let r1 =  pointercrate_test::demonlist::add_simple_record(100, player1.id, demon1, RecordStatus::Approved, connection).await;
+    let r2 =  pointercrate_test::demonlist::add_simple_record(70, player1.id, demon2, RecordStatus::Rejected, connection).await;
+    let r3 =  pointercrate_test::demonlist::add_simple_record(100, player2.id, demon2, RecordStatus::Rejected, connection).await;
 
     (player1.id, r1, r2, r3)
 }
 
 #[sqlx::test(migrations = "../migrations")]
 async fn unauthed_submit_for_player_with_locked_submission(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
-    let user = setup::add_normal_user(&mut connection).await;
+    let user =  pointercrate_test::user::add_normal_user(&mut connection).await;
     let player1 = DatabasePlayer::by_name_or_create("stardust1971", &mut connection).await.unwrap();
-    let demon1 = setup::add_demon("Bloodbath", 1, 87, player1.id, player1.id, &mut connection).await;
+    let demon1 =  pointercrate_test::demonlist::add_demon("Bloodbath", 1, 87, player1.id, player1.id, &mut connection).await;
 
-    setup::put_claim(user.inner().id, player1.id, true, true, &mut connection).await;
+    pointercrate_test::demonlist::put_claim(user.inner().id, player1.id, true, true, &mut connection).await;
 
     let submission =
         serde_json::json! {{"progress": 100, "demon": demon1, "player": "stardust1971", "video": "https://youtube.com/watch?v=1234567890"}};
@@ -121,11 +119,11 @@ async fn unauthed_submit_for_player_with_locked_submission(pool: Pool<Postgres>)
 
 #[sqlx::test(migrations = "../migrations")]
 async fn submit_existing_record(pool: Pool<Postgres>) {
-    let (clnt, mut connection) = setup::setup_rocket(pool).await;
+    let (clnt, mut connection) = pointercrate_test::demonlist::setup_rocket(pool).await;
 
     let player1 = DatabasePlayer::by_name_or_create("stardust1971", &mut connection).await.unwrap();
-    let demon1 = setup::add_demon("Bloodbath", 1, 50, player1.id, player1.id, &mut connection).await;
-    let existing = setup::add_simple_record(70, player1.id, demon1, RecordStatus::Approved, &mut connection).await;
+    let demon1 =  pointercrate_test::demonlist::add_demon("Bloodbath", 1, 50, player1.id, player1.id, &mut connection).await;
+    let existing =  pointercrate_test::demonlist::add_simple_record(70, player1.id, demon1, RecordStatus::Approved, &mut connection).await;
 
     let submission =
         serde_json::json! {{"progress": 60, "demon": demon1, "player": "stardust1971", "video": "https://youtube.com/watch?v=1234567890"}};
