+++ pointercrate-test/tests/it/mod.rs
@@ -1,0 +1,1 @@
+mod demonlist;
