+++ pointercrate-test/tests/main.rs
@@ -1,0 +1,1 @@
+mod it;
